buffer: Move buffer to service.rs and layer to layer.rs
